// src/components/App/Hack/Types.tsx
export enum RESULT {
  ELIGIBLE = 'ELIGIBLE',
  NOT_ELIGIBLE = 'NOT_ELIGIBLE',
}

export type TNotification = Record<'content' | 'emoji', string>

export type UserData = { key: string; value: string }